move handleGoogle to helper file as separate function
