+++ src/screens/signIn/index.js
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import Design from "./design";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import { login, loginWith3rdParty } from "./api";
+import { login } from "./api";
 import { validateForm } from "./validator/utils";
 
 function SignIn() {
@@ -53,45 +53,12 @@ function SignIn() {
     }
   };
 
-  const handleGoogle = (decoded) => {
-    const { given_name, family_name, email, sub } = decoded;
-    const payload = {
-      first_name: given_name,
-      last_name: family_name ? family_name : " ",
-      email: email,
-      third_party_user_id: sub,
-      third_party_type: "Google",
-    };
-    loginWith3rdParty(payload)
-      .then((response) => {
-        if (response.data.success) {
-          window.location.href = "/";
-        } else {
-          toast.success(response.data.message, {
-            position: "bottom-left",
-            autoClose: 2500,
-            hideProgressBar: true,
-            closeOnClick: false,
-            pauseOnHover: false,
-            draggable: false,
-            progress: undefined,
-            theme: "light",
-            type: response.data.success ? "success" : "error",
-          });
-        }
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
   return (
     <div>
       <Design
         userData={userData}
         setUserData={setUserData}
         handleSubmit={handleSubmit}
-        handleGoogle={handleGoogle}
       />
       <ToastContainer />
     </div>
