+++ src/screens/signUp/index.js
@@ -2,7 +2,7 @@
 import React, { useState } from "react";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import { addNewUser, loginWith3rdParty } from "./api";
+import { addNewUser } from "./api";
 import { validateForm } from "./validator/utils";
 import Design from "./design";
 
@@ -63,44 +63,12 @@ function SignUp() {
     }
   };
 
-  const handleGoogle = (decoded) => {
-    const { given_name, family_name, email, sub } = decoded;
-    const payload = {
-      first_name: given_name,
-      last_name: family_name ? family_name : " ",
-      email: email,
-      third_party_user_id: sub,
-      third_party_type: "Google",
-    };
-    loginWith3rdParty(payload)
-      .then((response) => {
-        if (response.data.success) {
-          window.location.href = "/";
-        } else {
-          toast.success(response.data.message, {
-            position: "bottom-left",
-            autoClose: 2500,
-            hideProgressBar: true,
-            closeOnClick: false,
-            pauseOnHover: false,
-            draggable: false,
-            progress: undefined,
-            theme: "light",
-            type: response.data.success ? "success" : "error",
-          });
-        }
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
   return (
     <div>
       <Design
         userData={userData}
         handleChange={handleChange}
         handleSubmit={handleSubmit}
-        handleGoogle={handleGoogle}
       />
       <ToastContainer />
     </div>
